requirements.txt
django~=4.1.5
django-rest-framework
djangorestframework~=3.14.0
django-cors-headers>=3.13.0

pandas
numpy

ghasedakpack
requests
